pyproject.toml
[tool.poetry]
name = "Poster"
version = "0.1.0"
description = "Minimal twitter like app"
authors = ["gabriel-viviani <gabrielviviani98@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
mypy = "^0.942"
flake8 = "^4.0.1"
SQLAlchemy = "^1.4.35"
fastapi-pagination = "^0.9.3"
psycopg2 = "^2.9.3"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
Faker = "^13.6.0"
pytest = "^7.1.2"
SQLAlchemy-Utils = "^0.38.2"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
